Refined vehicle behaviour
Suspicious vehicle checks are now only done on driver (as that’s where
the checks apply).

Vehicle checks include whether the vehicle is open or covered, whether
the unit has a weapon (civ) and if it is raised, and which uniform the
unit is wearing.

If the vehicle is covered or is a tank, racial profile checks are
ignored.

Defined variable for vehicle rather than using (vehicle _unit) each
time.