Overhaul of vehicle detection system
Vehicle detect while in incog or undercover is based on the visibility
and weirdness of each unit in that vehicle. If one unit in the vehicle
is acting really strangely, the rest will become compromised. Certain
vehicles are easily hidden in, others are pretty much hostile if the
crew isn't properly dressed. Takes into account vehicle position and
position visibility.