Fixed units switching to civilian
Workaround for weird engine behaviour of occasionally switching captive units to the civilian side under certain (difficult to pin down) scenarios. This manifested as enemies occasionally completely ignoring undercover units regardless of what they did. Now, the script detects if you are the wrong side to the undercover side and rectifies it with (untested) group switching. Biproduct means you can effectively place an undercover unit in the editor as any side at all and they will be automatically switched to the side defined as the undercover unit side in the setup.sqf, along with their group. This might be useful for making the perfectly disguised unit by simply placing an enemy down in the editor.