Added new options for civilian equipment additions
Added options for rare and super rare weapons (with configurable percentage chances).

Added option for civilians to spawn with vests from the civilian vest array.

Added ability for civilians to carry all weapons openly and not try to keep them concealed.

Added ability for weighting to be chosen when calling the addWeapon function.